requirements.txt
flask==2.3.2
opencv-python==4.7.0.72
numpy==1.24.3
mediapipe==0.10.0
python-dotenv==1.0.0
openai==0.27.8
